+++ app/app_components/userModule/controllers/githubOauthCtrl.js
@@ -1,8 +1,8 @@
 'use strict';
 angular.module('userModule')
     .controller('githubOauthCtrl',
-    ['$scope', '$state', '$stateParams', '$cookieStore', 'userService', 'STATE_DASHBOARD',
-        function ($scope, $state, $stateParams, $cookieStore, userService, STATE_DASHBOARD) {
+    ['$scope', '$state', '$stateParams', '$cookieStore', 'userService', 'notificationService', 'STATE_DASHBOARD',
+        function ($scope, $state, $stateParams, $cookieStore, userService, notificationService, STATE_DASHBOARD) {
             var cookieState = $cookieStore.get('state');
             $cookieStore.remove('state');
 
@@ -14,12 +14,12 @@ angular.module('userModule')
                 $scope.errorState = false;
                 userService.authenticateWithGithubAndLogInUser($stateParams.code).then(
                     function () {
+                        // start polling after login: cookie available --> no 403 errors
+                        notificationService.startPolling();
                         $state.go(STATE_DASHBOARD);
                     }, function () {
                         $scope.errorState = true;
                     }
                 );
-
             }
-
         }]);