+++ changelog.md
@@ -2,6 +2,7 @@
 ### Version 0.1.0-SNAPSHOT
 
  - [feature] add button design
+ - [feature] system notifications
  - [bug] close confirmation dialog completely
  - [feature] use a different github account for local dev development
  - [feature] PRT 99 added confirmation dialog, that asks to review older pull requests